+++ .appveyor.yml
@@ -14,9 +14,9 @@ install:
   - conda config --set always_yes yes --set changeps1 no
   - conda update -q conda
   - conda info -a
-  - conda create -q -n test-environment python=%PYTHON_VERSION% numpy scipy scikit-learn pandas matplotlib nose
+  - conda create -q -n test-environment python=%PYTHON_VERSION% numpy scipy scikit-learn pandas nose
   - activate test-environment
   - pip install nose-exclude
 
 test_script:
-  - nosetests -s -v --exclude-dir=mlxtend/tf_classifier --exclude-dir=mlxtend/tf_regressor --exclude-dir=mlxtend/tf_cluster --exclude-dir=mlxtend/general_plotting
+  - nosetests -s -v --exclude-dir=mlxtend/tf_classifier --exclude-dir=mlxtend/tf_regressor --exclude-dir=mlxtend/tf_cluster --exclude-dir=mlxtend/plotting
