+++ .travis.yml
@@ -1,10 +1,8 @@
-#sudo: required
-#dist: trusty
 language: python
 matrix:
     include:
         - python: 3.5
-          env: LATEST="false" TENSORFLOW="false" COVERAGE="false" NUMPY_VERSION="1.10.4" SCIPY_VERSION="0.17" SKLEARN_VERSION="0.17" PANDAS_VERSION="0.17.1" MATPLOTLIB_VERSION="1.5.1"
+          env: LATEST="false" TENSORFLOW="false" COVERAGE="false" NUMPY_VERSION="1.10.4" SCIPY_VERSION="0.17" SKLEARN_VERSION="0.17" PANDAS_VERSION="0.17.1"
         - python: 3.5
           env: LATEST="true" TENSORFLOW="false" COVERAGE="true"
         - python: 3.5
