+++ mlxtend/evaluate/__init__.py
@@ -4,11 +4,8 @@
 #
 # License: BSD 3 clause
 
-#from .decision_regions import plot_decision_regions
-from .learning_curves import plot_learning_curves
 from .scoring import scoring
 from .confusion_matrix import confusion_matrix
-from .plot_confusion_matrix import plot_confusion_matrix
 
-__all__ = ["plot_learning_curves",
-           "scoring", "confusion_matrix", "plot_confusion_matrix"]
+
+__all__ = ["scoring", "confusion_matrix"]
