+++ mlxtend/plotting/learning_curves.py
