+++ mlxtend/plotting/remove_chartjunk.py
