+++ mlxtend/plotting/scatter.py
