+++ mlxtend/plotting/stacked_barplot.py
